python/src/requirement.txt
numpy-1.25.5